renamed sample to polling
